requirements.txt
torch
transformers
datasets
trl
peft
pyyaml
accelerate
bitsandbytes
